+++ package/lib/src/widgets/mute_sound_button.dart
@@ -26,7 +26,7 @@ class MuteSoundButton extends StatelessWidget {
       }
 
       return PlayerButton(
-        size: responsive.ip(_.fullscreen.value ? 5 : 7),
+        size: responsive.buttonSize(),
         circle: false,
         backgroundColor: Colors.transparent,
         iconColor: Colors.white,